Add support for SAF
Use the normal (fast) file interface when possible, but add
support for SAF when necessary.

Also introduces an introduction to improve the first usage, but
it is also necessary for SAF.

Improves the permission requests. It only asks contacts/calendar
permission when enabling a collection, and also enables the
collection immediately when accapting.

Finally, it also shows collections unknown to DecSync but known
to Android.